refactor(timeline): 优化代码结构
