+++ components/timeline/timeline.tsx
@@ -1,103 +1,28 @@
 import {
     CSSProperties,
-    PropType,
     SlotsType,
     VNode,
     VNodeChild,
     computed,
     defineComponent,
-    ComponentObjectPropsOptions,
 } from 'vue';
 import { useTheme } from '../_theme/useTheme';
 import { timelineProps } from './props';
-import { COMPONENT_NAME, ICON_DEFAULT_COLOR, prefixCls } from './const';
+import { COMPONENT_NAME, prefixCls } from './const';
 import {
     calcDescPosition,
     calcTitlePosition,
     cls,
     getTitleOppositePosition,
-    isPresetIconTypes,
-    isValidRenderResult,
 } from './utils';
-import { useCustomIconRegister, useCustomIcons } from './useCustomIcons';
-import type { ComponentInnerProps } from './utilTypes';
+import { useCustomIcons } from './useCustomIcons';
+import Icon from './icon';
 import type {
     TimelineInnerProps as Props,
     TimelineSlots as Slots,
     TimelineUnboxSlots as UnboxSlots,
     TimelineNode,
 } from './props';
-
-const iconProps = {
-    index: { type: Number, required: true },
-    icon: { type: [String, Function] as PropType<TimelineNode['icon']> },
-    slotRender: { type: Function as PropType<UnboxSlots['icon']> },
-} as const satisfies ComponentObjectPropsOptions;
-const Icon = defineComponent({
-    name: `${COMPONENT_NAME}Icon`,
-    props: iconProps,
-    setup: (props: ComponentInnerProps<typeof iconProps>) => {
-        const customIcon = computed(() => {
-            let customIcon: VNodeChild;
-            // prop 的渲染函数优先级高于插槽
-            if (props.slotRender) {
-                customIcon = props.slotRender({ index: props.index });
-            } else if (typeof props.icon === 'function') {
-                customIcon = props.icon({ index: props.index });
-            }
-
-            // 自定义渲染没有内容时，fallback
-            if (!isValidRenderResult(customIcon)) {
-                customIcon = undefined;
-            }
-
-            return customIcon;
-        });
-
-        const isCustom = computed(() => !!customIcon.value);
-
-        const { iconRef } = useCustomIconRegister(props.index, isCustom);
-
-        return () => {
-            // 自定义图标
-            if (isCustom.value) {
-                return (
-                    <div
-                        ref={iconRef}
-                        class={[cls('item-icon'), cls('item-icon-custom')]}
-                    >
-                        {customIcon.value}
-                    </div>
-                );
-            }
-
-            // 自定义颜色
-            if (
-                typeof props.icon === 'string' &&
-                !isPresetIconTypes(props.icon)
-            ) {
-                return (
-                    <div
-                        ref={iconRef}
-                        class={cls('item-icon')}
-                        style={{ color: props.icon, borderColor: props.icon }}
-                    />
-                );
-            }
-
-            // 预设颜色
-            return (
-                <div
-                    ref={iconRef}
-                    class={[
-                        cls('item-icon'),
-                        cls(`item-icon-${props.icon ?? ICON_DEFAULT_COLOR}`),
-                    ]}
-                />
-            );
-        };
-    },
-});
 
 /** 渲染标题、辅助描述 */
 const renderNodeContent = ({
